Move soltype to a top-level sibling of solver
Relocate internal/solver/soltype/ to internal/soltype/, mirroring the
existing internal/type_system/ (representation) ↔ internal/checker/
(engine) sibling convention rather than nesting the representation inside
its consumer. Go gives subpackages no special relationship to their
parent, so this is a pure path/organization change; the one-way
solver → soltype dependency is unchanged.
