+++ internal/soltype/polarity.go
