+++ internal/soltype/type.go
