+++ internal/soltype/type_test.go
